catalogo/requirements.txt

C:\Users\ID13\Desktop\T\SENA\Curso Python\catalogo>python2.7.exe -m pip install
Cython==0.19.2
Django==1.5.7
Kivy==1.8.0
Markdown==2.6
Pygments==1.6
django-filter==0.9.2
djangorestframework==3.0.5
docutils==0.11
kivy-garden==0.1.1
plyer==1.1.2
pywin32==218
requests==2.2.1
